move 'Payload', 'PanicBuffer', and 'MAX_PAYLOAD_SIZE' to 'gear-core/buffer' module
