+++ core-backend/src/funcs.rs
@@ -36,14 +36,12 @@ use alloc::{format, string::String};
 use blake2::{digest::typenum::U32, Blake2b, Digest};
 use core::marker::PhantomData;
 use gear_core::{
-    buffer::{RuntimeBuffer, RuntimeBufferSizeError},
+    buffer::{Payload, PayloadSizeError, RuntimeBuffer, RuntimeBufferSizeError},
     costs::CostToken,
     env::DropPayloadLockBound,
     gas::CounterType,
     ids::{MessageId, ProgramId},
-    message::{
-        HandlePacket, InitPacket, MessageWaitedType, Payload, PayloadSizeError, ReplyPacket,
-    },
+    message::{HandlePacket, InitPacket, MessageWaitedType, ReplyPacket},
     pages::WasmPage,
 };
 use gear_core_errors::{MessageError, ReplyCode, SignalCode};
