+++ core/src/message/incoming.rs
@@ -17,9 +17,10 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
+    buffer::Payload,
     ids::{MessageId, ProgramId},
     message::{
-        common::MessageDetails, ContextStore, DispatchKind, GasLimit, Payload, StoredDispatch,
+        common::MessageDetails, ContextStore, DispatchKind, GasLimit, StoredDispatch,
         StoredMessage, Value,
     },
 };
