+++ core/src/message/init.rs
@@ -17,10 +17,11 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
+    buffer::Payload,
     ids::{prelude::*, CodeId, MessageId, ProgramId},
     message::{
-        Dispatch, DispatchKind, GasLimit, Message, Packet, Payload, Salt, StoredDispatch,
-        StoredMessage, Value,
+        Dispatch, DispatchKind, GasLimit, Message, Packet, Salt, StoredDispatch, StoredMessage,
+        Value,
     },
 };
 use scale_info::{
