+++ ethexe/runtime/common/src/state.rs
@@ -36,11 +36,10 @@ use core::{
 use ethexe_common::gear::{Message, Origin};
 pub use gear_core::program::ProgramState as InitStatus;
 use gear_core::{
+    buffer::Payload,
     ids::{prelude::MessageIdExt as _, ProgramId},
     memory::PageBuf,
-    message::{
-        ContextStore, DispatchKind, MessageDetails, Payload, ReplyDetails, StoredDispatch, Value,
-    },
+    message::{ContextStore, DispatchKind, MessageDetails, ReplyDetails, StoredDispatch, Value},
     pages::{numerated::tree::IntervalsTree, GearPage, WasmPage},
     program::MemoryInfix,
 };
