+++ gtest/src/log.rs
@@ -23,8 +23,9 @@ use crate::{
 };
 use core_processor::configs::BlockInfo;
 use gear_core::{
+    buffer::Payload,
     ids::{MessageId, ProgramId},
-    message::{Payload, StoredMessage, UserStoredMessage},
+    message::{StoredMessage, UserStoredMessage},
 };
 use gear_core_errors::{ErrorReplyReason, ReplyCode, SimpleExecutionError, SuccessReplyReason};
 use parity_scale_codec::{Codec, Encode};
