+++ pallets/gear/src/builtin.rs
@@ -18,10 +18,7 @@
 
 use super::*;
 use core_processor::common::JournalNote;
-use gear_core::{
-    ids::ProgramId,
-    message::{Payload, StoredDispatch},
-};
+use gear_core::{buffer::Payload, ids::ProgramId, message::StoredDispatch};
 
 /// Builtin actor `handle` function signature.
 pub type HandleFn<C, E> = dyn Fn(&StoredDispatch, &mut C) -> Result<Payload, E>;
